requirements.txt
# Dependances du projet
# Teste avec Python 3.8+ sur Google Colab
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
scikit-image>=0.19.0
tqdm>=4.64.0
scipy>=1.7.0
